# More optimizations