@jcrist 2017-03-28
cc @pinjutien, @yzard, @bansalr, this should provide some more speedups on top of #37.