@yzard 2017-03-28
@jcrist Salute!, thanks for your effort, we really appreciate, right now we will run on several hundred cores machines to see the results!