@jakirkham 2017-03-28
What size data (order of magnitude) are you pushing through this @yzard?